backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pillow==10.1.0
pdf2image==1.16.3
python-docx==1.1.0
trafilatura==1.6.1
pdfplumber==0.10.3
chromadb==0.4.22
sentence-transformers==2.2.2
langchain==0.0.350
python-dotenv==1.0.0
firebase-admin==6.2.0
openai==1.3.8